move mutation logic to parent component
