+++ src/app/dashboard/application-table.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import { useAtom } from "jotai";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../components/ui/table";
 import type { ApplicationQuestion, ApplicationWithResponses, RecruitmentCyclePhase } from "../types";
-import { api } from "~/trpc/react";
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuPortal, DropdownMenuSub, DropdownMenuSubContent, DropdownMenuSubTrigger, DropdownMenuTrigger } from "~/components/ui/dropdown-menu";
 import { Check, MoreVertical } from "lucide-react";
 import { Button } from "~/components/ui/button";
@@ -11,45 +9,20 @@ import ViewNotes from "./view-notes";
 import ApplicationDisplayDialog from "./application-display-dialog";
 import { FieldType } from "~/server/db/types";
 import FileViewerDialog from "~/components/ui/file-viewer-dialog";
-import { selectedRecruitmentCycleAtom } from "./atoms";
+import type { api } from "~/trpc/react";
+
 
 export default function ApplicationTable({
     displayedApplications,
     questions,
-    phases
+    phases,
+    setApplicationPhaseIdMutation
 }: {
     displayedApplications: ApplicationWithResponses[],
     questions: ApplicationQuestion[],
-    phases: RecruitmentCyclePhase[]
+    phases: RecruitmentCyclePhase[],
+    setApplicationPhaseIdMutation: ReturnType<typeof api.application.updatePhase.useMutation>
 }) {
-    const [cycleId] = useAtom(selectedRecruitmentCycleAtom);
-    const utils = api.useContext();
-    const setApplicationPhaseIdMutation = api.application.updatePhase.useMutation({
-        onMutate: async (update) => {
-            // cancel outgoing refetches that will overwrite data
-            await utils.application.getSubmittedApplicationsWithResponsesByCycleId.cancel(cycleId);
-            const previousApplications = utils.application.getSubmittedApplicationsWithResponsesByCycleId.getData(cycleId)!;
-
-            // optimistically update application phase
-            const updatedApplications = [...previousApplications];
-            const updatedApplicationIndex = updatedApplications.findIndex(a => a.id === update.applicationId);
-            updatedApplications[updatedApplicationIndex] = {
-                ...previousApplications.find(a => a.id === update.applicationId)!,
-                phaseId: update.phaseId
-            };
-            utils.application.getSubmittedApplicationsWithResponsesByCycleId.setData(
-                cycleId,
-                updatedApplications
-            );
-
-            return { previousApplications };
-        },
-        onError: (_err, _update, context) => {
-            utils.application.getSubmittedApplicationsWithResponsesByCycleId.setData(cycleId, context?.previousApplications);
-        },
-        onSettled: () => utils.application.getSubmittedApplicationsWithResponsesByCycleId.invalidate(cycleId)
-    });
-
     const setApplicationPhase = (applicationId: string, phaseId: string) => {
         const updatedApplication = displayedApplications.find(a => a.id === applicationId);
         if (!updatedApplication) throw new Error("Application not found");
