+++ src/app/dashboard/view-applications.tsx
@@ -63,6 +63,33 @@ export default function ViewApplications() {
     });
     const { data: phasesData, isLoading: phasesLoading } = api.recruitmentCyclePhase.getByCycleId.useQuery(cycleId);
 
+    const utils = api.useContext();
+    const setApplicationPhaseIdMutation = api.application.updatePhase.useMutation({
+        onMutate: async (update) => {
+            // cancel outgoing refetches that will overwrite data
+            await utils.application.getSubmittedApplicationsWithResponsesByCycleId.cancel(cycleId);
+            const previousApplications = utils.application.getSubmittedApplicationsWithResponsesByCycleId.getData(cycleId)!;
+
+            // optimistically update application phase
+            const updatedApplications = [...previousApplications];
+            const updatedApplicationIndex = updatedApplications.findIndex(a => a.id === update.applicationId);
+            updatedApplications[updatedApplicationIndex] = {
+                ...previousApplications.find(a => a.id === update.applicationId)!,
+                phaseId: update.phaseId
+            };
+            utils.application.getSubmittedApplicationsWithResponsesByCycleId.setData(
+                cycleId,
+                updatedApplications
+            );
+
+            return { previousApplications };
+        },
+        onError: (_err, _update, context) => {
+            utils.application.getSubmittedApplicationsWithResponsesByCycleId.setData(cycleId, context?.previousApplications);
+        },
+        onSettled: () => utils.application.getSubmittedApplicationsWithResponsesByCycleId.invalidate(cycleId)
+    });
+
     const copyEmails = () => navigator.clipboard.writeText(applicationsData?.map(a => a.email).join(",") ?? "");
     const copyNames = () => navigator.clipboard.writeText(applicationsData?.map(a => a.name).join(",") ?? "");
     const exportApplications = () => {
@@ -242,6 +269,7 @@ export default function ViewApplications() {
                             displayedApplications={applicationsData!}
                             questions={questionsData ?? []}
                             phases={phasesData ?? []}
+                            setApplicationPhaseIdMutation={setApplicationPhaseIdMutation}
                         />
                     </TabsContent>
                     <TabsContent value="board">
@@ -249,6 +277,7 @@ export default function ViewApplications() {
                             displayedApplications={applicationsData!}
                             questions={questionsData ?? []}
                             phases={phasesData ?? []}
+                            setApplicationPhaseIdMutation={setApplicationPhaseIdMutation}
                         />
                     </TabsContent>
                 </Tabs>
